requirements.txt
fastapi
uvicorn
openai
pypdf
python-docx
pydantic
